fix: changed the empty check for Binder object.

+++ diagram.go
@@ -2,7 +2,6 @@ package binder
 
 import (
 	"binder/db/model"
-	"log/slog"
 
 	"github.com/google/uuid"
 	"golang.org/x/xerrors"
@@ -10,7 +9,9 @@ import (
 
 func (b *Binder) EditDiagram(d *model.Diagram) (*model.Diagram, error) {
 
-	slog.Info("Call EditDiagram()")
+	if b == nil {
+		return nil, EmptyError
+	}
 
 	//新規指定だった場合
 	if d.Id == "" {
@@ -22,11 +23,11 @@ func (b *Binder) EditDiagram(d *model.Diagram) (*model.Diagram, error) {
 
 		err = b.fileSystem.CreateDiagramFile(d)
 		if err != nil {
-			return nil, xerrors.Errorf("fs.EditDiagram() error: %w", err)
+			return nil, xerrors.Errorf("fs.CreateDiagramFile() error: %w", err)
 		}
 		err = b.db.InsertDiagram(d, b.op)
 		if err != nil {
-			return nil, xerrors.Errorf("fs.InsertDiagram() error: %w", err)
+			return nil, xerrors.Errorf("db.InsertDiagram() error: %w", err)
 		}
 	} else {
 		err := b.db.UpdateDiagram(d, b.op)
@@ -40,28 +41,65 @@ func (b *Binder) EditDiagram(d *model.Diagram) (*model.Diagram, error) {
 	return d, nil
 }
 
+func (b *Binder) GetDiagram(id string) (*model.Diagram, error) {
+	if b == nil {
+		return nil, EmptyError
+	}
+	d, err := b.db.GetDiagram(id)
+	if err != nil {
+		return nil, xerrors.Errorf("db.GetDiagram() error: %w", err)
+	}
+	return d, nil
+}
+
+func (b *Binder) OpenDiagram(id string) ([]byte, error) {
+	if b == nil {
+		return nil, EmptyError
+	}
+	data, err := b.fileSystem.ReadDiagram(id)
+	if err != nil {
+		return nil, xerrors.Errorf("fs.ReadDiagram() error: %w", err)
+	}
+	return data, nil
+}
+
+func (b *Binder) SaveDiagram(id string, data []byte) error {
+
+	if b == nil {
+		return EmptyError
+	}
+	err := b.fileSystem.WriteDiagram(id, data)
+	if err != nil {
+		return xerrors.Errorf("fs.WriteDiagram() error: %w", err)
+	}
+
+	return nil
+}
+
 func (b *Binder) RemoveDiagram(id string) (*model.Diagram, error) {
 
+	if b == nil {
+		return nil, EmptyError
+	}
+
+	d, err := b.db.GetDiagram(id)
+	if err != nil {
+		return nil, xerrors.Errorf("db.GetDiagram() error: %w", err)
+	}
+
 	//ファイルを削除
-	err := b.fileSystem.DeleteDiagram(id)
+	err = b.fileSystem.DeleteDiagram(id)
 	if err != nil {
 		return nil, xerrors.Errorf("fs.DeleteDiagram() error: %w", err)
 	}
+
+	//DBを削除
 	err = b.db.DeleteDiagram(id)
 	if err != nil {
 		return nil, xerrors.Errorf("db.DeleteNote() error: %w", err)
 	}
-	return nil, nil
-}
 
-func (b *Binder) GetDiagram(id string) (*model.Diagram, error) {
-	return b.db.GetDiagram(id)
-}
+	//TODO コミット
 
-func (b *Binder) OpenDiagram(id string) ([]byte, error) {
-	return b.fileSystem.ReadDiagram(id)
-}
-
-func (b *Binder) SaveDiagram(id string, data []byte) error {
-	return b.fileSystem.WriteDiagram(id, data)
+	return d, nil
 }
